dbMemoからコマンドラインに関係する処理を抜き出す修正
