+++ 04.class/memo.js
@@ -2,6 +2,8 @@
 import minimist from "minimist";
 import readline from "readline";
 import dbMemo from "./dbMemo.js";
+import enquirer from "enquirer";
+const { Select } = enquirer;
 
 class MemoApp {
   async receiveUserInput() {
@@ -25,29 +27,65 @@ class MemoApp {
     const argv = minimist(process.argv.slice(2));
     const dm = new dbMemo();
     if (argv.l === undefined && argv.r === undefined && argv.d === undefined) {
-     this.createMemo(dm);
+      this.createMemo(dm);
     } else {
       //ここでDB関係を呼び出してオプションに合わせて出力したい
       if (argv.l) {
-        dm.listupMemo();
+        const row = await dm.listupMemo();
+        row.forEach((element) => console.log(element.title));
       }
       if (argv.r) {
-        dm.readMemo();
+        const row = await dm.readMemo();
+        const choices = row.map((memo) => ({
+          name: memo.title,
+          value: memo.content,
+        }));
+        const prompt = new Select({
+          name: "color",
+          message: "choose a note you want to see:",
+          choices: choices,
+          result() {
+            return this.focused.value;
+          },
+        });
+
+        prompt
+          .run()
+          .then((answer) => console.log(answer))
+          .catch(console.error);
       }
       if (argv.d) {
-        dm.deleteMemo();
+        const row = dm.deleteMemo();
+        const choices = row.map((memo) => ({
+          name: memo.title,
+          value: memo.id,
+        }));
+        const prompt = new Select({
+          name: "color",
+          message: "choose a note you want to delete:",
+          choices: choices,
+          result() {
+            return this.focused.value;
+          },
+        });
+
+        prompt
+          .run()
+          .then((answer) =>
+            this.db.run("DELETE FROM memo WHERE id == (?)", [answer]),
+          )
+          .catch(console.error);
       }
     }
   }
 
-  async createMemo(dm){
+  async createMemo(dm) {
     let userInput = await this.receiveUserInput();
     const title = userInput[0];
     const content = userInput.slice(0).join("\n");
     //ここでDB関係を呼び出して保存したい
     dm.inputMemo(title, content);
   }
-
 }
 
 async function main() {
